model/nfl/touchdown/nfltouchdowntrain/pyproject.toml
[tool.poetry]
name = "nfltouchdowntrain"
version = "0.1.0"
description = "Training model for NFL touchdowns"
authors = ["Christopher Mortimer <christopher@mortimer.xyz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
hydra-core = "^1.3.2"
python-dotenv = "^1.0.1"
google-cloud-aiplatform = "^1.74.0"
matplotlib = "^3.9.3"
joblib = "^1.4.2"
scikit-learn = "^1.6.1"
nfltouchdown = "^0.1.6"
# for testing changed features in nfltouchdown
#nfltouchdown = { path = "../nfltouchdown", develop = true }

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
ipykernel = "^6.29.5"
pytest-env = "^1.1.5"

[[tool.poetry.source]]
name = "vpy"
url = "https://australia-southeast1-python.pkg.dev/prj-xyz-shr-rep-0/rpo-vpy-0/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
